modularise and comment
